env: ti: Check reboot reason in BCB
*** PLEASE DO NOT MERGE.
*** This is only RFC, a discussion thread. Patch is only for the
*** reference here right now, to create a discussion context.

There are 2 possible reboot use-cases:
  1. User did "fastboot reboot bootloader" (we're setting "dofastboot"
     variable as an indicator in this case)
  2. User did "adb reboot bootloader" or "reboot bootloader" from Linux
     shell (Android sets "bootonce-bootloader" string into command field
     of BCB area inside of 'misc' partition as an indicator in this
     case)

We already had (1) handling implemented. This patch adds implementation
of (2). Possible drawbacks are:
  - user is able to re-define 'preboot' variable and break this
    mechanism this way
  - performance penalty due to scripting usage in 'preboot'

DISCUSSION ITEM 1: possible solution to those can be checking BCB area
(and 'dofastboot' variable) in C code, somewhere in board file, like
in late_init() function, and on;y set 'preboot' variable from there,
if conditions are met, to something like this:

    preboot=fastboot 1; setenv preboot;

so that next time 'preboot' variable will be empty. But this way also
has its shortcomings, like the necessity of board file modification,
need of BCB C API exposure, etc.

DISCUSSION ITEM 2: this patch only implements reboot reason handling for
TI platforms. I presume there could be another parties interested in
this feature. Should we somehow provide some generic way for handling
that (at least mechanism for checking), or is it a bad idea, and
everyone should handle this in their own way?

DISCUSSION ITEM 3: I didn't handle "reboot recovery" reason here. Does
anyone have some ideas on that? This doc [1] a bit of confuses me, but
as I understand, the 'recovery' partition will remain, so in case of
"reboot recovery" we just need to load recovery partition into the RAM
and run bootm for that. Is that correct or I'm missing something?

[1] https://source.android.com/devices/tech/ota/ab/ab_implement

Signed-off-by: Sam Protsenko <semen.protsenko@linaro.org>